refactor(java/racingcar/car): 패키지 정리
클래스 목적에 맞게 패키지 정리
- Car 관련 객체들을 racingcar/car 패키지로 이동
