+++ src/main/java/racingcar/car/CarName.java
@@ -1,4 +1,4 @@
-package racingcar;
+package racingcar.car;
 
 import utils.ErrorMessage;
